requirements.txt
pandas
numpy
streamlit
pillow
matplotlib
